+++ auth-service/src/main/java/com/ecommerce/auth/service/AdminService.java
@@ -1,0 +1,12 @@
+package com.ecommerce.auth.service;
+
+
+import com.ecommerce.auth.dto.UserDto;
+
+import java.util.List;
+
+public interface AdminService {
+    List<UserDto> getAllUsers();
+    UserDto getUserById(Long id);
+    void deleteUserById(Long id);
+}