+++ auth-service/src/main/java/com/ecommerce/auth/service/impl/AdminServiceImpl.java
@@ -1,0 +1,57 @@
+package com.ecommerce.auth.service.impl;
+
+
+import com.ecommerce.auth.dto.UserDto;
+import com.ecommerce.auth.exception.UserNotFoundException;
+import com.ecommerce.auth.exception.UserNotFoundException;
+import com.ecommerce.auth.model.User;
+import com.ecommerce.auth.repository.UserRepository;
+import com.ecommerce.auth.service.AdminService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+@RequiredArgsConstructor
+public class AdminServiceImpl implements AdminService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    public List<UserDto> getAllUsers() {
+        return userRepository.findAll().stream()
+                .map(this::convertToDto)
+                .collect(Collectors.toList());
+    }
+
+
+    @Override
+    public UserDto getUserById(Long id) {
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException(("User not found with id: " + id)));
+        return convertToDto(user);
+    }
+
+
+    @Override
+    public void deleteUserById(Long id) {
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException("User not found with id: " + id));
+        userRepository.delete(user);
+    }
+
+    private UserDto convertToDto(User user) {
+        return UserDto.builder()
+                .id(user.getId())
+                .name(user.getName())
+                .email(user.getEmail())
+                .role(user.getRole().name())
+                .build();
+    }
+
+
+}